Add files via upload
These are the files I've been trying to get to work, The Uninstall....bat is mine, the win10fix_full.bat is a batch I found and used to get windows 10 on my laptop so many years ago, and I liked the menus and checking for information. At this point I cannot get the if else statement in ::uninstall to work of the command fails out.